0.2.5

+++ src/thread.js
@@ -10,9 +10,16 @@ const __dirname = path.dirname(__filename)
 class Thread {
     #pathname = __dirname + '/worker.js'
 
-    constructor(fn, args, { autoStop } = {}) {
+    events = {
+        onStart: [],
+        onDone: [],
+        onStop: []
+    }
+
+    constructor(fn, args, callback, { autoStop } = {}) {
         this.fn = fn
         this.args = args
+        this.callback = callback
         this.autoStop = autoStop
     }
 
@@ -24,13 +31,28 @@ class Thread {
             worker.on('exit', code => this.#exit(code, reject))
             this.worker = worker
             Dispatcher.threads.push(this)
+            this.fireEvent('onStart')
         })
     }
 
     stop() {
+        if (this.worker === undefined) return
         this.worker.terminate()
         this.worker = undefined
         Dispatcher.threads.splice(Dispatcher.threads.find(thread => thread === this), 1)
+        this.fireEvent('onStop')
+    }
+
+    addEventListener(event, fn) {
+        this.events[event].push(fn)
+    }
+
+    removeEventListener(event, fn) {
+        this.events[event].splice(this.events[event].indexOf(fn), 1)
+    }
+
+    fireEvent(event, args) {
+        this.events[event].forEach(fn => fn(args))
     }
 
     #message(message, resolve) {
@@ -39,19 +61,24 @@ class Thread {
         if (this.autoStop !== undefined ? this.autoStop : Dispatcher.config.threads.autoStop)
             this.stop()
         resolve(message)
+        this.callback?.(message, null)
+        this.fireEvent('onDone')
     }
 
     #error(error, reject) {
         if (Dispatcher.config.logs.enabled)
             Dispatcher.config.logs.logger.error('[ THREADMAN THREAD ERROR ]', error)
         this.stop()
+        this.callback?.(null, error)
         reject(error)
     }
 
     #exit(code, reject) {
-        if (code === 0) return
         this.stop()
-        reject(new Error(`stopped with ${code} exit code`))
+        if (code === 0) return
+        const error = new Error(`stopped with ${code} exit code`)
+        this.callback?.(null, error)
+        reject(error)
     }
 }
